third_party/blink/web_tests/platform/generic/fast/dom/Range/surroundContents-for-invalid-state-error-expected.txt

PASS range.surroundContents(start) threw exception InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node..
PASS successfullyParsed is true

TEST COMPLETE
